src/requirements.txt
streamlit
pandas
sqlalchemy
requests
pyyaml